+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.0] - 2026-02-23
+
+### Added
+- ShaderGraph Organizer: Align context menu (Left, Right, Up, Down) for aligning selected nodes in the ShaderGraph editor.
+- ShaderGraph Organizer: Distribute context menu (Horizontal, Vertical) for evenly spacing selected nodes.
+
 ## [0.2.0] - 2026-01-27
 
 ### Added